+++ include/nc_runtime/pose2.h
@@ -18,4 +18,14 @@ public:
     Pose2(float x, float y, float ori_) : pos(x, y), ori(ori_) {}
 };
 
+forceinline bool operator==(Pose2 l, Pose2 r)
+{
+    return l.pos == r.pos && l.ori == r.ori;
+}
+
+forceinline bool operator!=(Pose2 l, Pose2 r)
+{
+    return l.pos != r.pos || l.ori != r.ori;
+}
+
 } // namespace nc